Updates namespace name
