+++ GJ2022/Entities/Pawns/Health/Bodyparts/Limbs/Human/HeadHuman.cs
@@ -11,8 +11,6 @@ namespace GJ2022.Entities.Pawns.Health.Bodyparts.Limbs.Human
     {
         public override BodySlots[] AllowedSlots => new BodySlots[] { BodySlots.SLOT_HEAD };
 
-        public override Organ[] ContainedOrgans => new Organ[];
-
         public override LimbFlags DefaultLimbFlags => LimbFlags.CRITICAL_LIMB | LimbFlags.NO_INSERTION;
     }
 }
