+++ GJ2022/Entities/Pawns/Health/Bodyparts/Limbs/Human/LegHuman.cs
@@ -6,7 +6,10 @@ using System.Threading.Tasks;
 
 namespace GJ2022.Entities.Pawns.Health.Bodyparts.Limbs.Human
 {
-    class LegHuman
+    public class LegHuman : Limb
     {
+        public override BodySlots[] AllowedSlots => new BodySlots[] { BodySlots.SLOT_LEG_LEFT, BodySlots.SLOT_LEG_RIGHT };
+
+        public override LimbFlags DefaultLimbFlags => LimbFlags.NONE;
     }
 }
