+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/BodyOrgans/Heart.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Body
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.BodyOrgans
 {
-    class Lung
+    class Heart
     {
     }
 }
