+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/BodyOrgans/Liver.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Body
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.BodyOrgans
 {
     class Liver
     {
