+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/HeadOrgans/Brain.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Head
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.HeadOrgans
 {
     public class Brain : Organ
     {
