+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/HeadOrgans/Ear.cs
@@ -4,15 +4,17 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Head
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.HeadOrgans
 {
-    public class Tongue : Organ
+    public class Ear : Organ
     {
 
-        public Tongue(Pawn parent) : base(parent)
+        public Ear(Pawn parent) : base(parent)
         { }
 
         public override OrganFlags DefaultOrganFlags => OrganFlags.NONE;
+
+        public override float HearingFactor => 50;
 
     }
 }
