+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/HeadOrgans/Eye.cs
@@ -4,15 +4,17 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Body
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.HeadOrgans
 {
-    public class Stomach : Organ
+    public class Eye : Organ
     {
 
-        public Stomach(Pawn parent) : base(parent)
+        public Eye(Pawn parent) : base(parent)
         { }
 
         public override OrganFlags DefaultOrganFlags => OrganFlags.NONE;
+
+        public override float VisionFactor => 50;
 
     }
 }
