+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/HeadOrgans/Nose.cs
@@ -4,17 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Head
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.HeadOrgans
 {
-    public class Eye : Organ
+    public class Nose : Organ
     {
-
-        public Eye(Pawn parent) : base(parent)
+        public Nose(Pawn parent) : base(parent)
         { }
 
         public override OrganFlags DefaultOrganFlags => OrganFlags.NONE;
-
-        public override float VisionFactor => 50;
-
     }
 }
