+++ GJ2022/Entities/Pawns/Health/Bodyparts/Organs/HeadOrgans/Tongue.cs
@@ -4,17 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.Head
+namespace GJ2022.Entities.Pawns.Health.Bodyparts.Organs.HeadOrgans
 {
-    public class Ear : Organ
+    public class Tongue : Organ
     {
 
-        public Ear(Pawn parent) : base(parent)
+        public Tongue(Pawn parent) : base(parent)
         { }
 
         public override OrganFlags DefaultOrganFlags => OrganFlags.NONE;
-
-        public override float HearingFactor => 50;
 
     }
 }
